improve ListName component
