+++ src/__tests__/components/List.test.tsx
@@ -11,21 +11,6 @@ describe('<List>', () => {
     store = storeFactory();
   });
 
-  test('should show `ListNameForm` when state of `isListFormVisible` is true', () => {
-    const { queryByTestId, getByTestId } = render(<List list={mockList} />, store);
-    fireEvent.click(getByTestId('listName'));
-
-    expect(queryByTestId('listName')).toBeNull();
-    expect(getByTestId('listNameForm')).toBeVisible();
-  });
-
-  test('should hide `ListNameForm` when state of `isListFormVisible` is false', () => {
-    const { queryByTestId, getByTestId } = render(<List list={mockList} />, store);
-
-    expect(queryByTestId('listNameForm')).toBeNull();
-    expect(getByTestId('listName')).toBeVisible();
-  });
-
   test('should show `ListMenu` when state of `isListMenuVisible` is true', () => {
     const { getByTestId } = render(<List list={mockList} />, store);
     fireEvent.click(getByTestId('listMenuButton'));
