+++ src/__tests__/components/ListName.test.tsx
@@ -2,28 +2,59 @@ import React from 'react';
 
 import { render, fireEvent, storeFactory } from '../../testUtils';
 import { Store } from '../../store';
-import { ListNameForm } from '../../components/ListNameForm';
+import { ListName } from '../../components/ListName';
 
-describe('<ListNameForm>', () => {
+describe('ListName component', () => {
   const listId = 1;
   const initialListName = 'initial list';
   let store: Store;
+  let updateList: jest.Mock;
+
   beforeEach(() => {
     store = storeFactory();
+    updateList = jest.fn();
   });
 
-  test('update state of `listName` upon list name text field changed', () => {
-    const setListFormVisible = jest.fn();
-    const updateList = jest.fn();
-    const { getByTestId } = render(
-      <ListNameForm
+  test('should show `ListNameForm` when state of `isListFormVisible` is true', () => {
+    const { queryByTestId, getByTestId } = render(
+      <ListName
         listId={listId}
         initialListName={initialListName}
-        setListFormVisible={setListFormVisible}
         updateList={updateList}
       />,
       store,
     );
+    fireEvent.click(getByTestId('listName'));
+
+    expect(queryByTestId('listName')).toBeNull();
+    expect(getByTestId('listNameForm')).toBeVisible();
+  });
+
+  test('should hide `ListNameForm` when state of `isListFormVisible` is false', () => {
+    const { queryByTestId, getByTestId } = render(
+      <ListName
+        listId={listId}
+        initialListName={initialListName}
+        updateList={updateList}
+      />,
+      store,
+    );
+
+    expect(queryByTestId('listNameForm')).toBeNull();
+    expect(getByTestId('listName')).toBeVisible();
+  });
+
+  test('update state of `listName` upon list name text field changed', () => {
+    const { getByTestId } = render(
+      <ListName
+        listId={listId}
+        initialListName={initialListName}
+        updateList={updateList}
+      />,
+      store,
+    );
+
+    fireEvent.click(getByTestId('listName'));
 
     const mockText = 'updated list';
     const listNameForm = getByTestId('listNameForm') as HTMLInputElement;
@@ -33,68 +64,62 @@ describe('<ListNameForm>', () => {
   });
 
   test('should not call `updateList` if a text field is blank when focus out from `listNameForm`', () => {
-    const setListFormVisible = jest.fn();
-    const updateList = jest.fn();
     const { getByTestId } = render(
-      <ListNameForm
+      <ListName
         listId={listId}
         initialListName={initialListName}
-        setListFormVisible={setListFormVisible}
         updateList={updateList}
       />,
       store,
     );
+
+    fireEvent.click(getByTestId('listName'));
 
     const mockText = '';
     const listNameForm = getByTestId('listNameForm') as HTMLInputElement;
     fireEvent.change(listNameForm, { target: { value: mockText } });
 
     fireEvent.blur(getByTestId('listNameForm'));
-    expect(setListFormVisible).toBeCalledWith(false);
     expect(updateList).not.toBeCalled();
   });
 
   test('should not call `updateList` if state of `listName` to equal `initialListName` when focus out from `listNameForm`', () => {
-    const setListFormVisible = jest.fn();
-    const updateList = jest.fn();
     const { getByTestId } = render(
-      <ListNameForm
+      <ListName
         listId={listId}
         initialListName={initialListName}
-        setListFormVisible={setListFormVisible}
         updateList={updateList}
       />,
       store,
     );
+
+    fireEvent.click(getByTestId('listName'));
 
     const mockText = initialListName;
     const listNameForm = getByTestId('listNameForm') as HTMLInputElement;
     fireEvent.change(listNameForm, { target: { value: mockText } });
 
     fireEvent.blur(getByTestId('listNameForm'));
-    expect(setListFormVisible).toBeCalledWith(false);
     expect(updateList).not.toBeCalled();
   });
 
   test('should call `updateList` if state of `listName` has changed when focus out from `listNameForm`', () => {
-    const setListFormVisible = jest.fn();
-    const updateList = jest.fn();
     const { getByTestId } = render(
-      <ListNameForm
+      <ListName
         listId={listId}
         initialListName={initialListName}
-        setListFormVisible={setListFormVisible}
         updateList={updateList}
       />,
       store,
     );
+
+    fireEvent.click(getByTestId('listName'));
 
     const mockText = 'updated list';
     const listNameForm = getByTestId('listNameForm') as HTMLInputElement;
     fireEvent.change(listNameForm, { target: { value: mockText } });
 
     fireEvent.blur(getByTestId('listNameForm'));
-    expect(setListFormVisible).toBeCalledWith(false);
     expect(updateList).toBeCalled();
   });
 });
