+++ src/css/List.scss
@@ -20,12 +20,6 @@
   margin-bottom: 5px;
   padding: 5px 5px;
   position: relative;
-  &__title {
-    cursor: pointer;
-    font-weight: bold;
-    flex: 1;
-    margin-right: 10px;
-  }
   &__icon {
     cursor: pointer;
     padding: 0 3px;
