+++ src/css/ListName.scss
@@ -5,3 +5,10 @@
   height: 26px;
   margin-right: 10px;
 }
+
+.listName {
+  cursor: pointer;
+  font-weight: bold;
+  flex: 1;
+  margin-right: 10px;
+}
